Adds a cancellable dialog that instructs the user to log in.
This fixes an issue where a thread could be leaked if the user does not log in and closes the browser.  Now, the user must at least click "cancel" which will stop the local server and cancel the login operation.
It has a dual purpose of instructing the user that we are waiting for him to log in -- in case the browser opened in a different monitor.

This dialog will show regardless of how login is invoked.

Change-Id: Ic5e411bfcb339e58541a92cff35d06bdba031b2b