Fix misaligned UI component in New Appengine Module Wizard
The problem is that HyperlinkLabel isn't smart enough to
wrap multiple lines of text, so we can't really use that.
JLabels can wrap text but when you give it HTML text with
a hyperlink the only way to make it clickable is to set
an action listener on the entire body of text.

The solution is to use a non-editable JEditorPane with
its background removed and its text size/font set to
match the rest of the surrounding UI. It seems kind of
dumb but I wasn't able to find any other nice solution
(and it it seems like other module wizards use this
approach too, for example AppEngineModuleWizard.java,
which is where I got most of this code from).

Before: https://screenshot.googleplex.com/X1CYDWeX3f
After: https://screenshot.googleplex.com/zX4bWPP3Li

Bug: 16823222

Change-Id: Ide5f7d999a44bc7e80219c6886274cbf3186690b