Invalid Full Method Name Uniqueness Validation error.
The Full Method Name Uniqueness Validation inspection shows a duplicate error message when a method with @ApiMethod annotation does not provide an name attribute. This is because the default name attribute value in the @Apimethod annotation is an empty string. In this case, the name attribute for this method should be ignore because a default method name will be provided by the App Engine SDK.

https://b.corp.google.com/issue?id=17291839&query=&cookieId=2014241081301089&disableB2Redirect=false&gsessionid=udYEa9A2A3w

Change-Id: I37dc196560c165a0f09775cccf5dc32be751379e