Integration test for Add Cloud Module
We've recently been broken externally by
changes that excluded our templates from
Add->New->Module because it set the context to null
even when a project was passed in.

This checkin adds a simple check to at least
ensure our template exists when it should and doesn't
when it should not.

Change-Id: I7834cee7062f71d5f8b99648a19e9d1ced7f7793