Cloud Repository Import Fails on Mac with Multiple GCP accounts
When using GCP repositories on Mac, the osxkeychain credential helper
is installed by default which caches credentials on its own per
domain.

This causes issues when a user attempts to manage two projects
with different GCP credentials.

To fix this, we can change the Mac credential helper config to
cache credentials based on full url instead of domain name.
We will only do this if we are setting up the Git repo.  If the
user has created the git repo on his own, he needs to manage the
config.

Change-Id: I8cb0fa8499e68267e1bfe2e8a1f3a2efb8b8f298