Cherry-pick from github (cloud-tools-for-intellij)
+ add in module dependency for login in gct-as

Usage Tracking for the Cloud Tools family of plugins

Keyed Extension Point that selects a single tracker based on the
platform prefix. If multiple implementations with the same key
exist, it will pick one of them (no way to specify which).

Extension points are used to allow keying and to allow platform
specific plugins to add tracker implementations

Change-Id: I1f60782ac01e5e5f7d649db1f96cf2287e52b8eb