build with gradle-intellij-plugin
I've updated how to use this new setup for importing into IntelliJ. Due
to an issue with how intellij does import you unfortunately have to
revert the imports changes to .idea after import to get the
preconfigured runconfiguration. There's also a couple ugly hacks to support
multi-plugin projects which will be fixed when
 https://github.com/JetBrains/gradle-intellij-plugin/issues/17 is fixed
in a few days.