User must select Cloud SDK home path, not /bin
Prior to this change, when configuring the Google Cloud SDK for
App Engine flexible environment, the user must select the bin
subdirectory of the SDK installation path.
This is counter to the expectation of the users that selecting
the SDK home directory should work.

Also, changing to store SDK home path rather than gcloud bin path
internally.