Add a CHANGELOG.md to our project
Our project is going to make our meaningful changes visible to users
using this CHANGELOG.md file.  All changes that could be interesting to
the user should be included in the [Unreleased] section as part of the
PR that makes the change.

I've seeded the file with changes from 0.9.4-beta to use as an example.

Closes #667.